requirements.txt
discord.py
watchdog
